rdm-api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
PyYAML==6.0.2
pydantic==2.9.2
